+++ backend/src/jobs/reconciliation/tasks/reconcile/buildSessions.js
@@ -2,7 +2,7 @@ const moment = require('moment');
 const computeScore = require('../../../../common/utils/computeScore');
 const { flatten } = require('../../../job-utils');
 const convertCommentToAvis = require('../../../../common/utils/convertCommentToAvis');
-const getReconcilatedAvis = require('./utils/getReconcilatedAvis');
+const filterAvisReconciliables = require('./rules/filterAvisReconciliables');
 
 module.exports = (intercarif, comments) => {
 
@@ -12,7 +12,7 @@ module.exports = (intercarif, comments) => {
             return acc;
         }
 
-        let reconciliated = getReconcilatedAvis(action, comments);
+        let reconciliated = filterAvisReconciliables(action, comments);
 
         return [
             ...acc,
